+++ source/extensions/lightspeed.trex.control.stagecraft/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.5.6]
+## Fixed
+- Added validation when opening a project with dependencies not in the project directory
+
 ## [1.5.5]
 ## Fixed
 - Fixed tests for new vertical tabs layout